+++ 3DPrimitiveDrawing/src/SUI/SUIManager.h
@@ -40,8 +40,8 @@ public:
 	static SUIManager* GetInstance();
 	static void DeleteInstance();
 
-	void Setup(int windowWidth, int windowHeight); //int argc, char** argv);
-	void SetWindowSize(int windowWidth, int windowHeight);
+	void Setup(float windowWidth, float windowHeight); //int argc, char** argv);
+	void SetWindowSize(float windowWidth, float windowHeight);
 
 	void Run();
 	float GetFPS();
